Exception handling moved in BaseServlet class. Rates type changed from Double to BigDecimal. ExchangeService created. All calculations moved from ExchangeServlet to ExchangeService.

+++ src/main/java/ru/klokov/dao/IExchangeRateDAO.java
@@ -4,6 +4,7 @@ import ru.klokov.exception.DatabaseException;
 import ru.klokov.model.Currency;
 import ru.klokov.model.ExchangeRate;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -11,5 +12,5 @@ public interface IExchangeRateDAO {
     List<ExchangeRate> findAll() throws DatabaseException;
     Optional<ExchangeRate> findByCurrencyPair(Currency baseCurrency, Currency targetCurrency) throws DatabaseException;
     ExchangeRate save(ExchangeRate exchangeRate) throws DatabaseException;
-    ExchangeRate update(ExchangeRate exchangeRate, double newRate) throws DatabaseException;
+    ExchangeRate update(ExchangeRate exchangeRate, BigDecimal newRate) throws DatabaseException;
 }
